refactor(frontend): extract useGanttContextMenu hook from GanttChart
Bundle the Gantt chart's context-menu concerns — target detection, open/close
position state, the task/group open handlers, and the action list for the
current target — into a useGanttContextMenu hook. The page passes its
navigation/edit callbacks in and consumes contextMenuState, closeContextMenu,
handleTaskContextMenu, handleGroupContextMenu, and contextMenuActions.
GanttChart drops the shouldOpenCustomContextMenu, suppressNativeContextMenu,
useContextMenuPositionState, and builder imports it no longer uses directly.
No behavior change.
